Replace service with WorkManager
